Rename webpack-helper -> webpack.
